backend/claude-engine/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Anthropic Claude SDK
anthropic>=0.40.0

# Google Gemini SDK
google-generativeai>=0.8.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Trading & Social Media APIs (for Unified Data API)
ccxt>=4.3.0,<5.0.0
praw==7.7.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Web Scraping (for Ground News)
playwright>=1.40.0

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6

# Caching
redis==5.0.1
hiredis==2.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Automated Prediction System
apscheduler==3.10.4
